TableChart/requirements.txt
Flask==2.1.2
mysql-connector-python==8.0.27
pandas==1.3.4
plotly==5.4.0